% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babysit.R
\name{babysit}
\alias{babysit}
\alias{state_init}
\title{An automatic routine to care for your virtual pet}
\usage{
state_init()
}
\arguments{
\item{tama}{An object of class Tama}

\item{end}{A POSIXct date: when should the baybsitting end ?}

\item{disc}{Is the babysitter allowed to scold the Tamagotchi when needed ?}
}
\description{
An automatic routine to care for your virtual pet
}
\examples{
guizmo = Tama()
guizmo$start()
babysit(guizmo, end = Sys.time() + 3600)
go(guizmo)
}
