% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click.r
\name{click}
\alias{click}
\title{Activate tamagotchi buttons}
\usage{
click(obj, button = c("A", "B", "C"), delay = 0.25)
}
\arguments{
\item{obj}{An object of class Tama}

\item{button}{A button to click, must be "A", "B" or "C" ; or a vector of several buttons for simultaneous clicking}

\item{delay}{The time to keep the buttons clicked, default is 0.25 second}
}
\description{
Activate tamagotchi buttons
}
