% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rom.R
\name{convert_rom}
\alias{convert_rom}
\title{Convert the rom.bin file into a rom.h base 12 format}
\usage{
convert_rom(bin, hfile = "rom.h", secret = T, ...)
}
\arguments{
\item{bin}{A binary ROM}

\item{hfile}{The output file name}

\item{secret}{Let a new (but familiar) secret character sneak in the game}
}
\description{
Convert the rom.bin file into a rom.h base 12 format
}
