% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2.R
\name{p2}
\alias{p2}
\title{Convert the sprites to gen2}
\usage{
p2(tama)
}
\arguments{
\item{tama}{An object of class Tama}
}
\description{
Convert the sprites to gen2
}
\examples{
guizmo = Tama()
p2(guizmo)
guizmo$start()
go(guizmo)
}
