% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.r
\name{Save}
\alias{Save}
\title{Save and load the game state}
\usage{
Save(obj, file)
}
\arguments{
\item{obj}{An object of class Tama}

\item{file}{The name of the file to save to or to load from}
}
\description{
Save and load the game state
}
\examples{
Save(tama,"myTama.txt")
Load(tama,"myTama.txt")
}
