% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{go}
\alias{go}
\title{A Shiny application to play with your virtual pet}
\usage{
go(tama, background = NULL, port = 1996, host = "127.0.0.1", light = TRUE)
}
\arguments{
\item{tama}{An object of class Tama}

\item{background}{A png image to use as background. It should be square and imported using png::readPNG}

\item{port}{The port to use for shiny, default is 1996}

\item{host}{The host to use for shiny, default is "127.0.0.1" (localhost)}

\item{light}{Should a lighter version of the app be used ? This is more adapted for non-web deployment. Default is TRUE}
}
\description{
A Shiny application to play with your virtual pet
}
\examples{
guizmo = Tama()
guizmo$start()
go(guizmo, light = T)
}
