% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamaR-exports.r
\name{Tama}
\alias{Tama}
\title{Create a Tama object from the C++ class}
\value{
A tamagotchi
}
\description{
Create a Tama object from the C++ class
}
\examples{
tama = new(Tama)
tama$GetMatrix()
tama$GetIcon()

tama$SetButton(0,T)
tama$GetButton()
tama$SetButton(0,F)

}
