% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamaR-exports.r
\name{Tama}
\alias{Tama}
\title{Instanciate a Tama object}
\value{
A tamagotchi
}
\description{
Instanciate a Tama object
}
\examples{
guizmo = Tama()
guizmo$display()
guizmo$click("B"); Sys.sleep(3)
for(i in 1:7) {guizmo$click("A"); Sys.sleep(.25)}
guizmo$display()
guizmo$save("tmp.txt")
guizmo$load("tmp.txt")

}
